Rename sql_table_test
