+++ test/storage/sql_table_test.cpp
@@ -1,10 +1,10 @@
-#include "storage/sql_table.h"
 #include <algorithm>
 #include <cstring>
 #include <random>
 #include <string>
 #include <utility>
 #include <vector>
+#include "storage/sql_table.h"
 #include "transaction/transaction_manager.h"
 #include "util/test_harness.h"
 #include "util/transaction_test_util.h"
@@ -13,14 +13,15 @@ namespace terrier {
 /**
  * Help class to simplify operations on a SqlTable
  */
-class SqlTableRW {
+class SqlTableTestRW {
  public:
-  explicit SqlTableRW(catalog::table_oid_t table_oid) : version_(0), table_oid_(table_oid) {}
-  ~SqlTableRW() {
+  explicit SqlTableTestRW(catalog::table_oid_t table_oid) : version_(0), table_oid_(table_oid) {}
+  ~SqlTableTestRW() {
     delete pri_;
     delete pr_map_;
     delete schema_;
     delete table_;
+    delete layout_;
   }
 
   /**
@@ -35,6 +36,32 @@ class SqlTableRW {
     cols_.emplace_back(name, type, nullable, oid);
   }
 
+  void AddColumn(transaction::TransactionContext *txn, std::string name, type::TypeId type, bool nullable,
+                 catalog::col_oid_t oid) {
+    // update columns, schema and layout
+    cols_.emplace_back(name, type, nullable, oid);
+    catalog::Schema new_schema(cols_);
+    delete schema_;
+    delete layout_;
+    schema_ = new catalog::Schema(cols_);
+    layout_ = new storage::BlockLayout(storage::StorageUtil::BlockLayoutFromSchema(*schema_).first);
+
+    table_->ChangeSchema(txn, new_schema);
+
+    col_oids_.clear();
+    for (const auto &c : cols_) {
+      col_oids_.emplace_back(c.GetOid());
+      LOG_INFO("{}", !c.GetOid());
+    }
+
+    delete pri_;
+    delete pr_map_;
+
+    // save information needed for (later) reading and writing
+    auto row_pair = table_->InitializerForProjectedRow(col_oids_, version_);
+    pri_ = new storage::ProjectedRowInitializer(std::get<0>(row_pair));
+    pr_map_ = new storage::ProjectionMap(std::get<1>(row_pair));
+  }
   /**
    * Create the SQL table.
    */
@@ -42,37 +69,63 @@ class SqlTableRW {
     schema_ = new catalog::Schema(cols_);
     table_ = new storage::SqlTable(&block_store_, *schema_, table_oid_);
 
+    layout_ = new storage::BlockLayout(storage::StorageUtil::BlockLayoutFromSchema(*schema_).first);
     for (const auto &c : cols_) {
       col_oids_.emplace_back(c.GetOid());
     }
 
     // save information needed for (later) reading and writing
-    auto row_pair = table_->InitializerForProjectedRow(col_oids_, storage::layout_version_t(0));
+    auto row_pair = table_->InitializerForProjectedRow(col_oids_, version_);
     pri_ = new storage::ProjectedRowInitializer(std::get<0>(row_pair));
     pr_map_ = new storage::ProjectionMap(std::get<1>(row_pair));
   }
 
   /**
-   * First step in writing a row.
+   * First step in inserting a row.
    */
-  void StartRow() {
-    insert_buffer_ = common::AllocationUtil::AllocateAligned(pri_->ProjectedRowSize());
-    insert_ = pri_->InitializeRow(insert_buffer_);
-  }
+  void StartInsertRow() { ResetProjectedRow(col_oids_); }
 
   /**
    * Insert the row into the table
    * @return slot where the row was created
    */
-  storage::TupleSlot EndRowAndInsert() {
-    auto txn = txn_manager_.BeginTransaction();
-    auto slot = table_->Insert(txn, *insert_, version_);
-    insert_ = nullptr;
-    txn_manager_.Commit(txn, TestCallbacks::EmptyCallback, nullptr);
-
-    delete[] insert_buffer_;
-    delete txn;
+  storage::TupleSlot EndInsertRow(transaction::TransactionContext *txn) {
+    auto slot = table_->Insert(txn, *pr_, version_);
+    pr_ = nullptr;
+    delete[] buffer_;
     return storage::TupleSlot(slot.GetBlock(), slot.GetOffset());
+  }
+
+  void StartUpdateRow(const std::vector<catalog::col_oid_t> &col_oids) {
+    auto pr_pair = table_->InitializerForProjectedRow(col_oids, version_);
+    buffer_ = common::AllocationUtil::AllocateAligned(pr_pair.first.ProjectedRowSize());
+    pr_ = pr_pair.first.InitializeRow(buffer_);
+
+    delete pri_;
+    delete pr_map_;
+    pri_ = new storage::ProjectedRowInitializer(std::get<0>(pr_pair));
+    pr_map_ = new storage::ProjectionMap(std::get<1>(pr_pair));
+  }
+
+  std::pair<bool, storage::TupleSlot> EndUpdateRow(transaction::TransactionContext *txn, storage::TupleSlot slot) {
+    auto result_pair = table_->Update(txn, slot, *pr_, *pr_map_, version_);
+    pr_ = nullptr;
+    delete[] buffer_;
+    return result_pair;
+  }
+
+  /**
+   * Check if a tuple is visible to you
+   * @param txn
+   * @param slot
+   * @return
+   */
+  bool Visible(transaction::TransactionContext *txn, storage::TupleSlot slot) {
+    buffer_ = common::AllocationUtil::AllocateAligned(pri_->ProjectedRowSize());
+    pr_ = pri_->InitializeRow(buffer_);
+    bool result = table_->Select(txn, slot, pr_, *pr_map_, version_);
+    delete[] buffer_;
+    return result;
   }
 
   /**
@@ -81,27 +134,28 @@ class SqlTableRW {
    * @param slot - tuple to read from
    * @return integer value
    */
-  uint32_t GetIntColInRow(int32_t col_num, storage::TupleSlot slot) {
-    auto txn = txn_manager_.BeginTransaction();
+  uint32_t GetIntColInRow(transaction::TransactionContext *txn, catalog::col_oid_t col_oid, storage::TupleSlot slot) {
     auto read_buffer = common::AllocationUtil::AllocateAligned(pri_->ProjectedRowSize());
     storage::ProjectedRow *read = pri_->InitializeRow(read_buffer);
     table_->Select(txn, slot, read, *pr_map_, version_);
-    byte *col_p = read->AccessForceNotNull(pr_map_->at(col_oids_[col_num]));
-    txn_manager_.Commit(txn, TestCallbacks::EmptyCallback, nullptr);
-    auto ret_val = *(reinterpret_cast<uint32_t *>(col_p));
-
-    delete txn;
+    byte *col_p = read->AccessWithNullCheck(pr_map_->at(col_oid));
+    uint32_t ret_val;
+    if (col_p == nullptr) {
+      ret_val = 12345;
+    } else {
+      ret_val = *(reinterpret_cast<uint32_t *>(col_p));
+    }
     delete[] read_buffer;
     return ret_val;
   }
 
   /**
-   * Save an integer, for insertion by EndRowAndInsert
+   * Save an integer, for insertion by EndInsertRow
    * @param col_num column number in the schema
    * @param value to save
    */
-  void SetIntColInRow(int32_t col_num, int32_t value) {
-    byte *col_p = insert_->AccessForceNotNull(pr_map_->at(col_oids_[col_num]));
+  void SetIntColInRow(catalog::col_oid_t col_oid, int32_t value) {
+    byte *col_p = pr_->AccessForceNotNull(pr_map_->at(col_oid));
     (*reinterpret_cast<uint32_t *>(col_p)) = value;
   }
 
@@ -112,12 +166,11 @@ class SqlTableRW {
    * @return malloc'ed C string (with null terminator). Caller must
    *   free.
    */
-  char *GetVarcharColInRow(int32_t col_num, storage::TupleSlot slot) {
-    auto txn = txn_manager_.BeginTransaction();
+  char *GetVarcharColInRow(transaction::TransactionContext *txn, catalog::col_oid_t(col_oid), storage::TupleSlot slot) {
     auto read_buffer = common::AllocationUtil::AllocateAligned(pri_->ProjectedRowSize());
     storage::ProjectedRow *read = pri_->InitializeRow(read_buffer);
     table_->Select(txn, slot, read, *pr_map_, version_);
-    byte *col_p = read->AccessForceNotNull(pr_map_->at(col_oids_[col_num]));
+    byte *col_p = read->AccessForceNotNull(pr_map_->at(col_oid));
 
     auto *entry = reinterpret_cast<storage::VarlenEntry *>(col_p);
     // stored string has no null terminator, add space for it
@@ -127,19 +180,18 @@ class SqlTableRW {
     std::memcpy(ret_st, entry->Content(), size);
     // add the null terminator
     *(ret_st + size - 1) = 0;
-    txn_manager_.Commit(txn, TestCallbacks::EmptyCallback, nullptr);
-    delete txn;
     delete[] read_buffer;
     return ret_st;
   }
 
+  storage::BlockLayout *GetLayout() { return layout_; }
   /**
    * Save a string, for insertion by EndRowAndInsert
    * @param col_num column number in the schema
    * @param st C string to save.
    */
-  void SetVarcharColInRow(int32_t col_num, const char *st) {
-    byte *col_p = insert_->AccessForceNotNull(pr_map_->at(col_oids_[col_num]));
+  void SetVarcharColInRow(catalog::col_oid_t col_oid, const char *st) {
+    byte *col_p = pr_->AccessForceNotNull(pr_map_->at(col_oid));
     // string size, without null terminator
     auto size = static_cast<uint32_t>(strlen(st));
     if (size <= storage::VarlenEntry::InlineThreshold()) {
@@ -153,79 +205,406 @@ class SqlTableRW {
     }
   }
 
- private:
-  storage::RecordBufferSegmentPool buffer_pool_{100, 100};
-  transaction::TransactionManager txn_manager_ = {&buffer_pool_, true, LOGGING_DISABLED};
-
+ public:
+  // This is a public field that transactions can set and read.
+  // The purpose is to record the version for each transaction. In reality this information should be retrieved from
+  // catalog. We just make it for the test case.
   storage::layout_version_t version_;
-
-  storage::BlockStore block_store_{100, 100};
-  catalog::table_oid_t table_oid_;
   storage::SqlTable *table_ = nullptr;
 
+ private:
+  storage::BlockStore block_store_{100, 100};
+  catalog::table_oid_t table_oid_;
+
+  // keep meta-data for most recent schema
   catalog::Schema *schema_ = nullptr;
   std::vector<catalog::Schema::Column> cols_;
   std::vector<catalog::col_oid_t> col_oids_;
+  storage::BlockLayout *layout_ = nullptr;
 
   storage::ProjectedRowInitializer *pri_ = nullptr;
   storage::ProjectionMap *pr_map_ = nullptr;
 
-  byte *insert_buffer_ = nullptr;
-  storage::ProjectedRow *insert_ = nullptr;
+  byte *buffer_ = nullptr;
+  storage::ProjectedRow *pr_ = nullptr;
+
+  void ResetProjectedRow(const std::vector<catalog::col_oid_t> &col_oids) {
+    auto pr_pair = table_->InitializerForProjectedRow(col_oids, version_);
+    buffer_ = common::AllocationUtil::AllocateAligned(pr_pair.first.ProjectedRowSize());
+    pr_ = pr_pair.first.InitializeRow(buffer_);
+
+    delete pri_;
+    delete pr_map_;
+    pri_ = new storage::ProjectedRowInitializer(std::get<0>(pr_pair));
+    pr_map_ = new storage::ProjectionMap(std::get<1>(pr_pair));
+  }
 };
 
 struct SqlTableTests : public TerrierTest {
   void SetUp() override { TerrierTest::SetUp(); }
 
   void TearDown() override { TerrierTest::TearDown(); }
+
+  storage::RecordBufferSegmentPool buffer_pool_{100, 100};
+  transaction::TransactionManager txn_manager_ = {&buffer_pool_, true, LOGGING_DISABLED};
 };
 
 // NOLINTNEXTLINE
-TEST_F(SqlTableTests, SelectInsertTest) {
-  SqlTableRW table(catalog::table_oid_t(2));
-
+TEST_F(SqlTableTests, SelectTest) {
+  SqlTableTestRW table(catalog::table_oid_t(2));
+  auto txn = txn_manager_.BeginTransaction();
   table.DefineColumn("id", type::TypeId::INTEGER, false, catalog::col_oid_t(0));
   table.DefineColumn("datname", type::TypeId::INTEGER, false, catalog::col_oid_t(1));
   table.Create();
-  table.StartRow();
-  table.SetIntColInRow(0, 100);
-  table.SetIntColInRow(1, 15721);
-  storage::TupleSlot row1_slot = table.EndRowAndInsert();
+  table.StartInsertRow();
+  table.SetIntColInRow(catalog::col_oid_t(0), 100);
+  table.SetIntColInRow(catalog::col_oid_t(1), 10000);
+  storage::TupleSlot row1_slot = table.EndInsertRow(txn);
 
-  table.StartRow();
-  table.SetIntColInRow(0, 200);
-  table.SetIntColInRow(1, 25721);
-  storage::TupleSlot row2_slot = table.EndRowAndInsert();
+  table.StartInsertRow();
+  table.SetIntColInRow(catalog::col_oid_t(0), 200);
+  table.SetIntColInRow(catalog::col_oid_t(1), 10001);
+  storage::TupleSlot row2_slot = table.EndInsertRow(txn);
 
-  uint32_t id = table.GetIntColInRow(0, row1_slot);
+  uint32_t id = table.GetIntColInRow(txn, catalog::col_oid_t(0), row1_slot);
   EXPECT_EQ(100, id);
-  uint32_t datname = table.GetIntColInRow(1, row1_slot);
-  EXPECT_EQ(15721, datname);
+  uint32_t datname = table.GetIntColInRow(txn, catalog::col_oid_t(1), row1_slot);
+  EXPECT_EQ(10000, datname);
 
-  id = table.GetIntColInRow(0, row2_slot);
+  // manually set the version of the transaction to be 1
+  table.version_ = storage::layout_version_t(1);
+  table.AddColumn(txn, "new_col", type::TypeId::INTEGER, true, catalog::col_oid_t(2));
+
+  id = table.GetIntColInRow(txn, catalog::col_oid_t(0), row1_slot);
+  EXPECT_EQ(100, id);
+  datname = table.GetIntColInRow(txn, catalog::col_oid_t(1), row1_slot);
+  EXPECT_EQ(10000, datname);
+
+  id = table.GetIntColInRow(txn, catalog::col_oid_t(0), row2_slot);
   EXPECT_EQ(200, id);
-  datname = table.GetIntColInRow(1, row2_slot);
-  EXPECT_EQ(25721, datname);
+  datname = table.GetIntColInRow(txn, catalog::col_oid_t(1), row2_slot);
+  EXPECT_EQ(10001, datname);
+
+  uint32_t new_col = table.GetIntColInRow(txn, catalog::col_oid_t(2), row1_slot);
+  EXPECT_EQ(12345, new_col);
+
+  txn_manager_.Commit(txn, TestCallbacks::EmptyCallback, nullptr);
+  delete txn;
+}
+
+// NOLINTNEXTLINE
+TEST_F(SqlTableTests, InsertTest) {
+  SqlTableTestRW table(catalog::table_oid_t(2));
+  auto txn = txn_manager_.BeginTransaction();
+  table.DefineColumn("id", type::TypeId::INTEGER, false, catalog::col_oid_t(0));
+  table.DefineColumn("datname", type::TypeId::INTEGER, false, catalog::col_oid_t(1));
+  table.Create();
+  table.StartInsertRow();
+  table.SetIntColInRow(catalog::col_oid_t(0), 100);
+  table.SetIntColInRow(catalog::col_oid_t(1), 10000);
+  storage::TupleSlot row1_slot = table.EndInsertRow(txn);
+
+  uint32_t id = table.GetIntColInRow(txn, catalog::col_oid_t(0), row1_slot);
+  EXPECT_EQ(100, id);
+  uint32_t datname = table.GetIntColInRow(txn, catalog::col_oid_t(1), row1_slot);
+  EXPECT_EQ(10000, datname);
+
+  // manually set the version of the transaction to be 1
+  table.version_ = storage::layout_version_t(1);
+  table.AddColumn(txn, "new_col", type::TypeId::INTEGER, true, catalog::col_oid_t(2));
+
+  // insert (300, 10002, null)
+  table.StartInsertRow();
+  table.SetIntColInRow(catalog::col_oid_t(0), 300);
+  table.SetIntColInRow(catalog::col_oid_t(1), 10002);
+  storage::TupleSlot row3_slot = table.EndInsertRow(txn);
+
+  // insert (400, 10003, 42)
+  table.StartInsertRow();
+  table.SetIntColInRow(catalog::col_oid_t(0), 400);
+  table.SetIntColInRow(catalog::col_oid_t(1), 10003);
+  table.SetIntColInRow(catalog::col_oid_t(2), 42);
+  storage::TupleSlot row4_slot = table.EndInsertRow(txn);
+
+  id = table.GetIntColInRow(txn, catalog::col_oid_t(0), row3_slot);
+  EXPECT_EQ(300, id);
+  datname = table.GetIntColInRow(txn, catalog::col_oid_t(1), row3_slot);
+  EXPECT_EQ(10002, datname);
+  uint32_t new_col = table.GetIntColInRow(txn, catalog::col_oid_t(2), row3_slot);
+  EXPECT_EQ(12345, new_col);
+
+  id = table.GetIntColInRow(txn, catalog::col_oid_t(0), row4_slot);
+  EXPECT_EQ(400, id);
+  datname = table.GetIntColInRow(txn, catalog::col_oid_t(1), row4_slot);
+  EXPECT_EQ(10003, datname);
+  datname = table.GetIntColInRow(txn, catalog::col_oid_t(2), row4_slot);
+  EXPECT_EQ(42, datname);
+
+  txn_manager_.Commit(txn, TestCallbacks::EmptyCallback, nullptr);
+  delete txn;
+}
+
+// NOLINTNEXTLINE
+TEST_F(SqlTableTests, DeleteTest) {
+  SqlTableTestRW table(catalog::table_oid_t(2));
+  auto txn = txn_manager_.BeginTransaction();
+  table.DefineColumn("id", type::TypeId::INTEGER, false, catalog::col_oid_t(0));
+  table.DefineColumn("datname", type::TypeId::INTEGER, false, catalog::col_oid_t(1));
+  table.Create();
+
+  // insert (100, 10000)
+  table.StartInsertRow();
+  table.SetIntColInRow(catalog::col_oid_t(0), 100);
+  table.SetIntColInRow(catalog::col_oid_t(1), 10000);
+  storage::TupleSlot row1_slot = table.EndInsertRow(txn);
+
+  // insert (200, 10001)
+  table.StartInsertRow();
+  table.SetIntColInRow(catalog::col_oid_t(0), 100);
+  table.SetIntColInRow(catalog::col_oid_t(1), 10001);
+  storage::TupleSlot row2_slot = table.EndInsertRow(txn);
+
+  // manually set the version of the transaction to be 1
+  table.version_ = storage::layout_version_t(1);
+  table.AddColumn(txn, "new_col", type::TypeId::INTEGER, true, catalog::col_oid_t(2));
+
+  // insert (300, 10002, null)
+  table.StartInsertRow();
+  table.SetIntColInRow(catalog::col_oid_t(0), 300);
+  table.SetIntColInRow(catalog::col_oid_t(1), 10002);
+  storage::TupleSlot row3_slot = table.EndInsertRow(txn);
+
+  // insert (400, 10003, 42)
+  table.StartInsertRow();
+  table.SetIntColInRow(catalog::col_oid_t(0), 400);
+  table.SetIntColInRow(catalog::col_oid_t(1), 10003);
+  table.SetIntColInRow(catalog::col_oid_t(2), 42);
+  storage::TupleSlot row4_slot = table.EndInsertRow(txn);
+
+  // delete (100, 10000, null) and (300, 10002, null)
+  EXPECT_TRUE(table.table_->Delete(txn, row1_slot, table.version_));
+  EXPECT_TRUE(table.table_->Delete(txn, row3_slot, table.version_));
+
+  EXPECT_TRUE(table.Visible(txn, row2_slot));
+  EXPECT_TRUE(table.Visible(txn, row4_slot));
+
+  EXPECT_FALSE(table.Visible(txn, row1_slot));
+  EXPECT_FALSE(table.Visible(txn, row3_slot));
+  txn_manager_.Commit(txn, TestCallbacks::EmptyCallback, nullptr);
+  delete txn;
+}
+
+// NOLINTNEXTLINE
+TEST_F(SqlTableTests, UpdateTest) {
+  SqlTableTestRW table(catalog::table_oid_t(2));
+  auto txn = txn_manager_.BeginTransaction();
+  table.DefineColumn("id", type::TypeId::INTEGER, false, catalog::col_oid_t(0));
+  table.DefineColumn("datname", type::TypeId::INTEGER, false, catalog::col_oid_t(1));
+  table.Create();
+
+  // insert (100, 10000)
+  table.StartInsertRow();
+  table.SetIntColInRow(catalog::col_oid_t(0), 100);
+  table.SetIntColInRow(catalog::col_oid_t(1), 10000);
+  storage::TupleSlot row1_slot = table.EndInsertRow(txn);
+
+  // insert (200, 10001)
+  table.StartInsertRow();
+  table.SetIntColInRow(catalog::col_oid_t(0), 200);
+  table.SetIntColInRow(catalog::col_oid_t(1), 10001);
+  storage::TupleSlot row2_slot = table.EndInsertRow(txn);
+
+  // manually set the version of the transaction to be 1
+  table.version_ = storage::layout_version_t(1);
+  table.AddColumn(txn, "new_col", type::TypeId::INTEGER, true, catalog::col_oid_t(2));
+
+  // insert (300, 10002, null)
+  table.StartInsertRow();
+  table.SetIntColInRow(catalog::col_oid_t(0), 300);
+  table.SetIntColInRow(catalog::col_oid_t(1), 10002);
+  table.EndInsertRow(txn);
+
+  // insert (400, 10003, 42)
+  table.StartInsertRow();
+  table.SetIntColInRow(catalog::col_oid_t(0), 400);
+  table.SetIntColInRow(catalog::col_oid_t(1), 10003);
+  table.SetIntColInRow(catalog::col_oid_t(2), 42);
+  storage::TupleSlot row4_slot = table.EndInsertRow(txn);
+
+  // update (200, 10001, null) -> (200, 11001, null)
+  std::vector<catalog::col_oid_t> update_oids;
+  update_oids.emplace_back(catalog::col_oid_t(1));
+  table.StartUpdateRow(update_oids);
+  table.SetIntColInRow(catalog::col_oid_t(1), 11001);
+  auto result = table.EndUpdateRow(txn, row2_slot);
+
+  EXPECT_TRUE(result.first);
+  EXPECT_EQ(result.second.GetBlock(), row2_slot.GetBlock());
+  EXPECT_EQ(result.second.GetOffset(), row2_slot.GetOffset());
+  uint32_t new_val = table.GetIntColInRow(txn, catalog::col_oid_t(1), row2_slot);
+  EXPECT_EQ(new_val, 11001);
+
+  // update (400, 10003, 42) -> (400, 11003, 420)
+  LOG_INFO("----------------------------")
+  update_oids.clear();
+  update_oids.emplace_back(catalog::col_oid_t(1));
+  update_oids.emplace_back(catalog::col_oid_t(2));
+  table.StartUpdateRow(update_oids);
+  table.SetIntColInRow(catalog::col_oid_t(1), 11003);
+  table.SetIntColInRow(catalog::col_oid_t(2), 420);
+  result = table.EndUpdateRow(txn, row4_slot);
+
+  EXPECT_TRUE(result.first);
+  EXPECT_EQ(result.second.GetBlock(), row4_slot.GetBlock());
+  EXPECT_EQ(result.second.GetOffset(), row4_slot.GetOffset());
+  new_val = table.GetIntColInRow(txn, catalog::col_oid_t(1), result.second);
+  EXPECT_EQ(new_val, 11003);
+  new_val = table.GetIntColInRow(txn, catalog::col_oid_t(2), result.second);
+  EXPECT_EQ(new_val, 420);
+
+  // update (100, 10000, null) -> (100, 11000, 420)
+  LOG_INFO("----------------------------")
+  update_oids.clear();
+  update_oids.emplace_back(catalog::col_oid_t(1));
+  update_oids.emplace_back(catalog::col_oid_t(2));
+  table.StartUpdateRow(update_oids);
+  table.SetIntColInRow(catalog::col_oid_t(1), 11000);
+  table.SetIntColInRow(catalog::col_oid_t(2), 420);
+  result = table.EndUpdateRow(txn, row1_slot);
+
+  EXPECT_TRUE(result.first);
+  EXPECT_NE(result.second.GetBlock(), row1_slot.GetBlock());
+  new_val = table.GetIntColInRow(txn, catalog::col_oid_t(1), result.second);
+  EXPECT_EQ(new_val, 11000);
+  new_val = table.GetIntColInRow(txn, catalog::col_oid_t(2), result.second);
+  EXPECT_EQ(new_val, 420);
+
+  txn_manager_.Commit(txn, TestCallbacks::EmptyCallback, nullptr);
+  delete txn;
+}
+
+// NOLINTNEXTLINE
+TEST_F(SqlTableTests, ScanTest) {
+  SqlTableTestRW table(catalog::table_oid_t(2));
+  auto txn = txn_manager_.BeginTransaction();
+  table.DefineColumn("id", type::TypeId::INTEGER, false, catalog::col_oid_t(0));
+  table.DefineColumn("datname", type::TypeId::INTEGER, false, catalog::col_oid_t(1));
+  table.Create();
+
+  // insert (100, 10000)
+  table.StartInsertRow();
+  table.SetIntColInRow(catalog::col_oid_t(0), 100);
+  table.SetIntColInRow(catalog::col_oid_t(1), 10000);
+  table.EndInsertRow(txn);
+
+  // insert (200, 10001)
+  table.StartInsertRow();
+  table.SetIntColInRow(catalog::col_oid_t(0), 200);
+  table.SetIntColInRow(catalog::col_oid_t(1), 10001);
+  table.EndInsertRow(txn);
+
+  // manually set the version of the transaction to be 1
+  table.version_ = storage::layout_version_t(1);
+  table.AddColumn(txn, "new_col", type::TypeId::INTEGER, true, catalog::col_oid_t(2));
+
+  // insert (300, 10002, null)
+  table.StartInsertRow();
+  table.SetIntColInRow(catalog::col_oid_t(0), 300);
+  table.SetIntColInRow(catalog::col_oid_t(1), 10002);
+  table.EndInsertRow(txn);
+
+  // insert (400, 10003, 42)
+  table.StartInsertRow();
+  table.SetIntColInRow(catalog::col_oid_t(0), 400);
+  table.SetIntColInRow(catalog::col_oid_t(1), 10003);
+  table.SetIntColInRow(catalog::col_oid_t(2), 42);
+  table.EndInsertRow(txn);
+
+  // begin scan
+  std::vector<catalog::col_oid_t> all_col_oids;
+  all_col_oids.emplace_back(0);
+  all_col_oids.emplace_back(1);
+  all_col_oids.emplace_back(2);
+
+  auto pc_pair = table.table_->InitializerForProjectedColumns(all_col_oids, 10, table.version_);
+  byte *buffer = common::AllocationUtil::AllocateAligned(pc_pair.first.ProjectedColumnsSize());
+  storage::ProjectedColumns *pc = pc_pair.first.Initialize(buffer);
+
+  // scan
+  auto start_pos = table.table_->begin();
+  table.table_->Scan(txn, &start_pos, pc, pc_pair.second, table.version_);
+
+  // check the number of tuples we found
+  EXPECT_EQ(pc->NumTuples(), 4);
+
+  // check the if we get (100, 10000, null)
+  auto row1 = pc->InterpretAsRow(*table.GetLayout(), 0);
+  byte *value = row1.AccessWithNullCheck(pc_pair.second.at(catalog::col_oid_t(0)));
+  EXPECT_NE(value, nullptr);
+  uint32_t id = *reinterpret_cast<uint32_t *>(value);
+  EXPECT_EQ(id, 100);
+  value = row1.AccessWithNullCheck(pc_pair.second.at(catalog::col_oid_t(1)));
+  EXPECT_NE(value, nullptr);
+  uint32_t datname = *reinterpret_cast<uint32_t *>(value);
+  EXPECT_EQ(datname, 10000);
+  value = row1.AccessWithNullCheck(pc_pair.second.at(catalog::col_oid_t(2)));
+  EXPECT_EQ(value, nullptr);
+
+  // check the if we get (200, 10001, null)
+  auto row2 = pc->InterpretAsRow(*table.GetLayout(), 1);
+  value = row2.AccessWithNullCheck(pc_pair.second.at(catalog::col_oid_t(0)));
+  EXPECT_NE(value, nullptr);
+  id = *reinterpret_cast<uint32_t *>(value);
+  EXPECT_EQ(id, 200);
+  value = row2.AccessWithNullCheck(pc_pair.second.at(catalog::col_oid_t(1)));
+  EXPECT_NE(value, nullptr);
+  datname = *reinterpret_cast<uint32_t *>(value);
+  EXPECT_EQ(datname, 10001);
+  value = row2.AccessWithNullCheck(pc_pair.second.at(catalog::col_oid_t(2)));
+  EXPECT_EQ(value, nullptr);
+
+  // check the if we get (400, 10003, 42)
+  auto row4 = pc->InterpretAsRow(*table.GetLayout(), 3);
+  value = row4.AccessWithNullCheck(pc_pair.second.at(catalog::col_oid_t(0)));
+  EXPECT_NE(value, nullptr);
+  id = *reinterpret_cast<uint32_t *>(value);
+  EXPECT_EQ(id, 400);
+  value = row4.AccessWithNullCheck(pc_pair.second.at(catalog::col_oid_t(1)));
+  EXPECT_NE(value, nullptr);
+  datname = *reinterpret_cast<uint32_t *>(value);
+  EXPECT_EQ(datname, 10003);
+  value = row4.AccessWithNullCheck(pc_pair.second.at(catalog::col_oid_t(2)));
+  EXPECT_NE(value, nullptr);
+  uint32_t new_col = *reinterpret_cast<uint32_t *>(value);
+  EXPECT_EQ(new_col, 42);
+
+  delete[] buffer;
+  txn_manager_.Commit(txn, TestCallbacks::EmptyCallback, nullptr);
+  delete txn;
 }
 
 // NOLINTNEXTLINE
 TEST_F(SqlTableTests, VarlenInsertTest) {
-  SqlTableRW table(catalog::table_oid_t(2));
+  SqlTableTestRW table(catalog::table_oid_t(2));
+  auto txn = txn_manager_.BeginTransaction();
 
   table.DefineColumn("id", type::TypeId::INTEGER, false, catalog::col_oid_t(0));
   table.DefineColumn("datname", type::TypeId::VARCHAR, false, catalog::col_oid_t(1));
   table.Create();
 
-  table.StartRow();
-  table.SetIntColInRow(0, 100);
-  table.SetVarcharColInRow(1, "name");
-  storage::TupleSlot row_slot = table.EndRowAndInsert();
+  table.StartInsertRow();
+  table.SetIntColInRow(catalog::col_oid_t(0), 100);
+  table.SetVarcharColInRow(catalog::col_oid_t(1), "name");
+  storage::TupleSlot row_slot = table.EndInsertRow(txn);
 
-  uint32_t id = table.GetIntColInRow(0, row_slot);
+  uint32_t id = table.GetIntColInRow(txn, catalog::col_oid_t(0), row_slot);
   EXPECT_EQ(100, id);
-  char *table_name = table.GetVarcharColInRow(1, row_slot);
+  char *table_name = table.GetVarcharColInRow(txn, catalog::col_oid_t(1), row_slot);
   EXPECT_STREQ("name", table_name);
   delete[] table_name;
+
+  txn_manager_.Commit(txn, TestCallbacks::EmptyCallback, nullptr);
+  delete txn;
 }
 
 }  // namespace terrier
